+++ lib/pb/serializer.rb
@@ -10,14 +10,58 @@ require "pb/serializer/oneof"
 module Pb
   module Serializer
     class Error < StandardError; end
+    class InvalidConfigurationError < Error; end
     class MissingMessageTypeError < Error; end
     class UnknownFieldError < Error; end
     class ValidationError < Error; end
     class ConflictOneofError < Error; end
-    class InvalidOptionError < Error; end
+    class InvalidAttributeOptionError < Error; end
     class MissingFieldError < Error; end
 
+    class Configuration
+      # @!attribute logger
+      #   @return [Logger]
+      attr_accessor :logger
+      # @!attribute [r] missing_field_behavior
+      #   @return [:raise, :warn, :ignore] default: `:raise`
+      attr_reader :missing_field_behavior
+
+      def initialize
+        self.missing_field_behavior = :raise
+        self.logger = Logger.new(STDOUT)
+      end
+
+      # @param v [:raise, :warn, :ignore]
+      def missing_field_behavior=(v)
+        @missing_field_behavior = v
+
+        unless %i(raise warn ignore).include?(v)
+          raise InvalidConfigurationError, "missing_field_behavior #{v} is not allowed"
+        end
+      end
+    end
+
     class << self
+      # @example
+      #   Pb::Serializer.configuration do |c|
+      #     c.missing_field_behavior = :raise  # :raise, :warn or :ignore (defualt: :raise)
+      #   end
+      # @yield [c]
+      # @yieldparam [Configuration] config
+      def configure
+        yield configuration
+      end
+
+      # @return [Pb::Serializer::Configuration]
+      def configuration
+        @configuraiton ||= Configuration.new
+      end
+
+      # @return [Logger]
+      def logger
+        configuration.logger
+      end
+
       # @param [Google::Protobuf::Descriptor]
       def build_default_mask(descriptor)
         set =